+++ frontend/src/views/AddAnimals/AddAnimalsModal.tsx
@@ -1,7 +1,7 @@
 /**
  * @summary This is the Add Animal list Tab
  */
-import React, { useState } from 'react';
+import React, { useContext, useEffect, useState } from 'react';
 import BeefCattle from './AnimalFormComponents/BeefCattle';
 import DairyCattle from './AnimalFormComponents/DairyCattle';
 import OtherAnimals from './AnimalFormComponents/OtherAnimals';
@@ -15,11 +15,14 @@ import {
   OTHER_ANIMAL_IDS,
   OtherAnimalData,
   OtherAnimalId,
+  SelectOption,
+  Animal,
 } from '@/types';
 import UnselectedAnimal from './AnimalFormComponents/UnselectedAnimal';
 import Modal, { ModalProps } from '@/components/common/Modal/Modal';
 import { INITIAL_BEEF_FORM_DATA, INITIAL_DAIRY_FORM_DATA } from '@/constants';
 import { INITIAL_POULTRY_FORM_DATA } from '@/constants/Animals';
+import { APICacheContext } from '@/context/APICacheContext';
 
 type AddAnimalsModalProps = {
   initialModalData: AnimalData | undefined;
@@ -36,6 +39,8 @@ export default function AddAnimalsModal({
   ...props
 }: AddAnimalsModalProps & Omit<ModalProps, 'title' | 'children' | 'onOpenChange'>) {
   const [formData, setFormData] = useState<AnimalData | undefined>(initialModalData);
+  const [animalOptions, setAnimalOptions] = useState<SelectOption[]>([]);
+  const apiCache = useContext(APICacheContext);
 
   const handleSubmit = (newFormData: AnimalData) => {
     if (rowEditIndex !== undefined) {
@@ -84,6 +89,20 @@ export default function AddAnimalsModal({
     });
   };
 
+  useEffect(() => {
+    apiCache.callEndpoint('/api/animals/').then((response: { status?: any; data: any }) => {
+      if (response.status === 200) {
+        const { data } = response;
+        const options = (data as Animal[]).map((row) => ({ id: row.id, label: row.name }));
+        // TODO: REMOVE ONCE WE HAVE SWINE
+        // This is a lazy way to take it out of the list
+        options.splice(options.length - 1, 1);
+        setAnimalOptions(options);
+      }
+    });
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, []);
+
   return (
     <Modal
       title="Add animals"
@@ -92,12 +111,14 @@ export default function AddAnimalsModal({
     >
       {formData === undefined && (
         <UnselectedAnimal
+          animalOptions={animalOptions}
           handleInputChanges={handleInputChanges}
           onCancel={onClose}
         />
       )}
       {formData?.animalId === BEEF_COW_ID && (
         <BeefCattle
+          animalOptions={animalOptions}
           formData={formData}
           handleInputChanges={handleInputChanges}
           handleSubmit={handleSubmit}
@@ -106,6 +127,7 @@ export default function AddAnimalsModal({
       )}
       {formData?.animalId === DAIRY_COW_ID && (
         <DairyCattle
+          animalOptions={animalOptions}
           formData={formData}
           handleInputChanges={handleInputChanges}
           handleSubmit={handleSubmit}
@@ -114,6 +136,7 @@ export default function AddAnimalsModal({
       )}
       {formData?.animalId === POULTRY_ID && (
         <Poultry
+          animalOptions={animalOptions}
           formData={formData}
           handleInputChanges={handleInputChanges}
           handleSubmit={handleSubmit}
@@ -122,6 +145,7 @@ export default function AddAnimalsModal({
       )}
       {formData !== undefined && OTHER_ANIMAL_IDS.some((id) => id === formData.animalId) && (
         <OtherAnimals
+          animalOptions={animalOptions}
           formData={formData as OtherAnimalData}
           handleInputChanges={handleInputChanges}
           handleSubmit={handleSubmit}
