+++ frontend/src/views/AddAnimals/AnimalFormComponents/UnselectedAnimal.tsx
@@ -1,18 +1,19 @@
+import { SelectOption } from '@/types';
 import AnimalFormWrapper from './AnimalFormWrapper';
 
 type UnselectedAnimalProps = {
+  animalOptions: SelectOption[];
   handleInputChanges: (changes: { [name: string]: string | number | undefined }) => void;
   onCancel: () => void;
 };
 
-export default function UnselectedAnimal({ handleInputChanges, onCancel }: UnselectedAnimalProps) {
+export default function UnselectedAnimal(props: UnselectedAnimalProps) {
   return (
     <AnimalFormWrapper
       selectedAnimalId={undefined}
-      handleInputChanges={handleInputChanges}
-      onCancel={onCancel}
       onSubmit={() => {}}
       isConfirmDisabled
+      {...props}
     >
       {null}
     </AnimalFormWrapper>
