+++ frontend/src/views/ManureAndImports/ManureAndImports.tsx
@@ -18,7 +18,7 @@ import {
   AnimalData,
   ManureType,
   Animal,
-  SelectOption,
+  DAIRY_COW_ID,
 } from '@/types';
 import {
   DefaultSolidManureConversionFactors,
@@ -50,7 +50,6 @@ export default function ManureAndImports() {
   const [animalList] = useState<Array<AnimalData>>(state.nmpFile.years[0]?.FarmAnimals || []);
 
   const [animals, setAnimals] = useState<Animal[]>([]);
-  const [subtypeList, setSubtypeList] = useState<SelectOption[]>([]);
   const [editMaterialName, setEditMaterialName] = useState<string | null>(null);
   const [manures, setManures] = useState<NMPFileImportedManureData[]>(
     state.nmpFile.years[0]?.ImportedManures || [],
@@ -65,7 +64,7 @@ export default function ManureAndImports() {
     useState<NMPFileImportedManureData>(DefaultManureFormData);
 
   const hasDairyCattle = useMemo(
-    () => animalList.some((animal) => animal.animalId === '2'),
+    () => animalList.some((animal) => animal.animalId === DAIRY_COW_ID),
     [animalList],
   );
 
@@ -188,16 +187,6 @@ export default function ManureAndImports() {
       if (response.status === 200) {
         const { data } = response;
         setAnimals(data);
-      }
-    });
-
-    apiCache.callEndpoint('api/animal_subtypes/').then((response) => {
-      if (response.status === 200) {
-        const { data } = response;
-        const subType: { id: string; label: string }[] = (
-          data as { id: number; name: string }[]
-        ).map((row) => ({ id: row.id.toString(), label: row.name }));
-        setSubtypeList((prev) => [...prev, ...subType]);
       }
     });
     // eslint-disable-next-line react-hooks/exhaustive-deps
