+++ frontend/src/views/NutrientAnalysis/NutrientAnalysis.tsx
@@ -10,7 +10,7 @@ import { DataGrid, GridColDef, GridRenderCellParams } from '@mui/x-data-grid';
 import { Button, Button as ButtonGov, ButtonGroup } from '@bcgov/design-system-react-components';
 import { AppTitle, PageTitle, ProgressStepper, Tabs } from '../../components/common';
 import { StyledContent } from './nutrientAnalysis.styles';
-import { AnimalData, NMPFileImportedManureData } from '@/types';
+import { AnimalData, DAIRY_COW_ID, NMPFileImportedManureData } from '@/types';
 import useAppState from '@/hooks/useAppState';
 import { MANURE_IMPORTS, FIELD_LIST, CALCULATE_NUTRIENTS, STORAGE } from '@/constants/routes';
 import { NMPFileFarmManureData } from '@/types/NMPFileFarmManureData';
@@ -41,7 +41,9 @@ export default function NutrientAnalysis() {
 
   const hasDairyCattle = useMemo(
     () =>
-      state.nmpFile.years[0]?.FarmAnimals?.some((animal: AnimalData) => animal.animalId === '2'),
+      state.nmpFile.years[0]?.FarmAnimals?.some(
+        (animal: AnimalData) => animal.animalId === DAIRY_COW_ID,
+      ),
     [state.nmpFile.years],
   );
